split expiration from timer
